Parser script
This is a parsing function.
It suppose to get list of strings as input and glue a 'rsync' command line string from that.
There still some work left. But at least it do something and could be useful.
This parser can  and will be modified according to our requirements and needs.